# Implement dict_jsonable