<!-- Thank you for your contribution! -->
<!-- Don't worry about making lots of commits on a pull request, they'll be squashed on merge anyway -->

## Change Summary

This PR implements a `dict_jsonable()` method that returns a `dict` (equivalent to the one returned by `dict()` ) but with all the values converted to JSON serialisable values (like it is done to generate the `str` returned by `json()` ).

The return of `dict_jsonable()` is a `dict` that can be passed to other functions that need something that can be serialised to JSON, like with `json.dumps()`.

For example, NoSQL DB connections or web frameworks.

<!-- Please give a short summary of the changes. -->

## Related issue number

I didn't create an issue with the feature request before the PR, should I?

<!-- Are there any issues opened that will be resolved by merging this change? -->

## Performance Changes

pydantic cares about performance, if there's any risk performance changed on this PR, 
please run `make benchmark-pydantic` before and after the change:

**No performance loss:**

* before:

```
                                pydantic time=1.047s, success=49.95%
                                pydantic time=0.945s, success=49.95%
                                pydantic time=0.943s, success=49.95%
                                pydantic time=0.874s, success=49.95%
                                pydantic time=0.920s, success=49.95%
                                pydantic best=0.874s, avg=0.946s, stdev=0.063s

                                pydantic best=29.143μs/iter avg=31.531μs/iter stdev=2.113μs/iter
```

* after:

```
                                pydantic time=0.871s, success=49.95%
                                pydantic time=0.928s, success=49.95%
                                pydantic time=0.909s, success=49.95%
                                pydantic time=0.941s, success=49.95%
                                pydantic time=0.906s, success=49.95%
                                pydantic best=0.871s, avg=0.911s, stdev=0.026s

                                pydantic best=29.047μs/iter avg=30.370μs/iter stdev=0.881μs/iter
```

## Checklist

* [x] Unit tests for the changes exist
* [x] Tests pass on CI and coverage remains at 100%
* [x] Documentation reflects the changes
* [x] No performance deterioration (if applicable)
* [x] `HISTORY.rst` has been updated
  * if this is the first change since a release, please add a new section
  * include the issue number or this pull request number `#<number>`
  * if you're not a regular contributer please include your github username `@whatever`