@samuelcolvin 2018-12-03
Humm, I'm not entirely opposed to this, but for me, it would only make sense if we could use the same method in the `json()` method, then we would have one interface to JSON.

Unfortunately, this method (even after rearranging slightly to make it 4 x faster and using ujson to encode) is still significantly slower than `json.dumps(data, default=pydantic_encoder)`, see [these](https://gist.github.com/samuelcolvin/1647958c99cc59873f36400048d77ae2) very crude benchmarks. (Let me know if I've done anything hideous to make it much slower than it should be, or used a daft data set)

Since:
1. Not everyone is using JSON and we shouldn't obsess over JSON
2. This method is slower
3. End users are likely to want to customise this kind of thing a lot
4. End users can implement this in a few lines themeselves (see below)

I'm inclined to refuse this. But I'll leave it open for a week or so to canvas opinions from others.

To implement in your own code, this should only take:

```py
from types import GeneratorType
from pydantic.json import pydantic_encoder

def jsonable_encoder(obj):
    if isinstance(obj, (str, int, float, type(None))):
        return obj
    if isinstance(obj, dict):
        return {jsonable_encoder(key): jsonable_encoder(value) for key, value in obj.items()}
    if isinstance(obj, (list, set, frozenset, GeneratorType, tuple)):
        return [jsonable_encoder(item) for item in obj]
    return pydantic_encoder(obj)
```