@tiangolo 2018-12-03
Get it. Thanks for checking it and for the example implementation.

---

I guess one option would be to just add that function from your comment to `utils.py`. 

But if it's the case that _End users are likely to want to customise this kind of thing a lot_, then it wouldn't make that much sense having that function in `utils.py`.

I'll use your suggestion and implement it on my own code assuming this will be rejected (unless something changes in the next days).